Refactor_1

+++ app/src/main/java/com/example/mvp_retrofit/entity/Activity.kt
@@ -8,10 +8,10 @@ data class Activity(
     @Json(name = "activity") val activity: String,
     @Json(name = "type") val type: String,
     @Json(name = "participants") val participants: Int,
-    @Json(name = "price") val price: Int,
+    @Json(name = "price") val price: Double,
     @Json(name = "link") val link: String,
     @Json(name = "key") val key: String,
-    @Json(name = "accessibility") val accessibility: Int
+    @Json(name = "accessibility") val accessibility: Double
 )
 
 @JsonClass(generateAdapter = true)
@@ -19,8 +19,8 @@ data class ActivityViewState(
     @Json(name = "activity") val activity: String,
     @Json(name = "type") val type: String,
     @Json(name = "participants") val participants: Int,
-    @Json(name = "price") val price: Int,
+    @Json(name = "price") val price: Double,
     @Json(name = "link") val link: String,
     @Json(name = "key") val key: String,
-    @Json(name = "accessibility") val accessibility: Int
+    @Json(name = "accessibility") val accessibility: Double
 )
